[Refactor] seperate all of SideBarType.ts #89
