+++ app/src/lib/common/types/ComponentTypes/Header/SideBar/SideBarIconType.ts
@@ -1,0 +1,8 @@
+export interface SideBarIconPropsType extends SideBarIconStyledPropsType {
+  iconSize?: string;
+  onClick: () => void;
+}
+
+export interface SideBarIconStyledPropsType {
+  iconMargin?: string;
+}
