+++ app/src/lib/common/types/ComponentTypes/Header/SideBar/SideBarItemsType.ts
@@ -1,0 +1,12 @@
+export interface SideBarItemsPropsType {
+  sideBarItems: any[];
+  setIsClickedSideBarIcon: (isClickedSideBarIcon: boolean) => void;
+}
+
+export interface SideBarItemsElementType {
+  title: string;
+}
+
+export interface SideBarItemsStyledPropsType {
+  margin?: string;
+}
