+++ app/src/lib/common/types/ComponentTypes/Header/SideContainerType.ts
@@ -1,0 +1,7 @@
+import { ChannelType } from '../ChannelType';
+import { SideBarOptionPropsType } from './SideBar/SideBarType';
+
+export interface SideContainerPropsType {
+  channels: ChannelType[];
+  sideBarOption: SideBarOptionPropsType;
+}
