+++ app/src/lib/components/Header/SideContainer/SideBarItems.tsx
@@ -1,24 +1,24 @@
 import styled from 'styled-components';
 import {
-  SideBarItemPropsType,
-  SideBarItemElementType,
-  SideBarItemStyledPropsType,
-} from '../../../common/types/ComponentTypes/Header/SideBarType';
+  SideBarItemsPropsType,
+  SideBarItemsElementType,
+  SideBarItemsStyledPropsType,
+} from '../../../common/types/ComponentTypes/Header/SideBar/SideBarItemsType';
 
 const ItemContainer = styled.div`
   color: ${({ color }) => color ?? 'black'};
   text-decoration: none;
 `;
 
-const Item = styled.div<SideBarItemStyledPropsType>`
+const Item = styled.div<SideBarItemsStyledPropsType>`
   cursor: pointer;
   margin: ${({ margin }) => margin ?? '10px 0px'};
 `;
 
-const SideBarItems = ({ sideBarItems, setIsClickedSideBarIcon }: SideBarItemPropsType) => {
+const SideBarItems = ({ sideBarItems, setIsClickedSideBarIcon }: SideBarItemsPropsType) => {
   return (
     <>
-      {sideBarItems.map(({ title }: SideBarItemElementType, idx) => (
+      {sideBarItems.map(({ title }: SideBarItemsElementType, idx) => (
         <ItemContainer
           key={idx}
           onClick={() => {
